backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1
hiredis==2.3.2

# OpenAI & AI
openai==1.10.0
langchain==0.1.6
langchain-openai==0.0.5
pinecone-client==3.0.1
tiktoken==0.5.2

# Sentiment Analysis
textblob==0.17.1
vaderSentiment==3.3.2

# WebSocket
websockets==12.0

# Background Tasks
celery==5.3.4
flower==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Email
python-email-validator==2.1.0
aiosmtplib==3.0.1
jinja2==3.1.3

# SMS/WhatsApp (Twilio)
twilio==8.11.0

# Validation & Utilities
email-validator==2.1.0
python-slugify==8.0.1

# Monitoring & Logging
sentry-sdk==1.40.0
python-json-logger==2.0.7

   # Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
